continuos fixes

+++ src/main/java/pddurfubot/cache/DataCacheInterface.java
@@ -3,7 +3,7 @@ package pddurfubot.cache;
 import pddurfubot.handlers.BotState;
 
 public interface DataCacheInterface {
-    void setUsersCurrentBotState(int userId, BotState botState);
+    void setUsersCurrentBotState(Long userId, BotState botState);
 
-    BotState getUsersCurrentBotState(int userId);
+    BotState getUsersCurrentBotState(Long userId);
 }
