+++ src/main/java/pddurfubot/cache/UserDataCache.java
@@ -5,15 +5,15 @@ import java.util.Map;
 import java.util.HashMap;
 
 public class UserDataCache implements DataCacheInterface {
-    private Map<Integer, BotState> userBotStates = new HashMap<>();
+    private Map<Long, BotState> userBotStates = new HashMap<>();
 
     @Override
-    public void setUsersCurrentBotState(int userId, BotState botState) {
+    public void setUsersCurrentBotState(Long userId, BotState botState) {
         userBotStates.put(userId, botState);
     }
 
     @Override
-    public BotState getUsersCurrentBotState(int userId) {
+    public BotState getUsersCurrentBotState(Long userId) {
         BotState botState = userBotStates.get(userId);
         if (botState == null) {
             botState = BotState.ANSWER_HELP;
