+++ src/main/java/pddurfubot/handlers/CommandBuild.java
@@ -3,10 +3,8 @@ package pddurfubot.handlers;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Message;
 
-import java.util.Arrays;
-
 public class CommandBuild {
-    public static SendMessage BuildOutputMessage(BotState botState, Message message, int userId) {
+    public static SendMessage BuildOutputMessage(BotState botState, Message message, Long userId) {
         return (
                 SendMessage.builder()
                 .chatId(message.getChatId().toString())
